beta version

+++ laundry/views.py
@@ -2,8 +2,9 @@ import json
 import re
 
 from django.contrib import messages
-from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth import authenticate, login, logout, get_user_model
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
 from django.http import HttpResponseRedirect, JsonResponse
 from django.shortcuts import render, redirect
 from django.urls import reverse
@@ -18,10 +19,9 @@ from .models import Laundry
 def login_view(request):
     if request.method == "POST":
 
-        # Attempt to sign user in
-        username = 'vitap_mh3'
         data = json.loads(request.body)
         password = _md5(data.get('password'))
+        username = data.get('username')
 
         user = authenticate(request, username=username, password=password)
         print(user, password, username)
@@ -34,7 +34,9 @@ def login_view(request):
     else:
         if request.user.is_authenticated:
             return HttpResponseRedirect(reverse("index"))
-        return render(request, "laundry/login.html")
+        return render(request, "laundry/login.html",{
+            "users": User.objects.filter(username__startswith='vitap_')
+        })
 
 
 def logout_view(request):
@@ -42,20 +44,68 @@ def logout_view(request):
     return HttpResponseRedirect(reverse("index"))
 
 
+def register_view(request):
+    if request.method == "POST":
+        data = json.loads(request.body)
+        password = _md5(data.get('password'))
+        username = data.get('username')
+
+        # Check if the username already exists
+        if User.objects.filter(username=username).exists():
+            return JsonResponse({'status': 'username_exists'})
+
+        # Create a new user
+        try:
+            user = User.objects.create_user(username=username, password=password)
+            user.save()
+            login(request, user)
+            return JsonResponse({'status': 'registered'})
+        except Exception as e:
+            print(e)
+            return JsonResponse({'status': 'error'})
+    else:
+        if request.user.is_authenticated:
+            return HttpResponseRedirect(reverse("index"))
+        return render(request, "laundry/register.html")
+
+
+def change_password_view(request):
+    if request.method == "POST":
+        data = json.loads(request.body)
+        username = data.get('username')
+        password = _md5(data.get('password'))
+        try:
+            user = get_user_model().objects.get(username=username)
+            user.set_password(password)
+            user.save()
+            login(request, user)
+            return JsonResponse({'status': 'password_changed'})
+
+        except get_user_model().DoesNotExist:
+            return JsonResponse({'status': 'user_not_found'})
+        except Exception as e:
+            print(e)
+            return JsonResponse({'status': 'error'})
+    else:
+        return render(request, "laundry/change_password.html",{
+            "users": User.objects.filter(username__startswith='vitap_')
+        })
+
+
 @login_required(login_url='login')
 def index(request):
     if not request.user.is_authenticated:
         return HttpResponseRedirect(reverse("login"))
-
+    laundries = Laundry.objects.filter(user=request.user)
+    latest_laundries = laundries.order_by('-date').order_by('reg_number').exclude(out=True)
     return render(request, "laundry/index.html", {
-        'laundries': Laundry.objects.filter(user=request.user)
+        'laundries': latest_laundries
     })
 
 
 def info(request, reg_number):
     reg_number = reg_number.upper().strip()
-    laundries = Laundry.objects.filter(reg_number=reg_number).order_by('-date')
-    print(laundries, reg_number)
+    laundries = Laundry.objects.filter(reg_number=reg_number, user=request.user).order_by('-date')
     return render(request, "laundry/info.html", {
         "reg_number": reg_number,
         "laundries": laundries
@@ -65,6 +115,9 @@ def info(request, reg_number):
 def delete(request, id):
     if request.method == "POST":
         laundry = Laundry.objects.get(id=id)
+        if laundry.user != request.user:
+            messages.error(request, "You are not authorized to delete this laundry.")
+            return redirect('index')
         reg = laundry.reg_number
         laundry.delete()
         messages.error(request, f"Deleted laundry {id}.")
@@ -85,6 +138,9 @@ def toggle(request, id):
         laundry = Laundry.objects.get(id=id)
         data = json.loads(request.body)
         status = data.get('status', None)
+        if laundry.user != request.user:
+            return JsonResponse({'success': False, 'error': 'You are not authorized to toggle this laundry.'},
+                                status=403)
 
         if status:
             laundry.out = True if status == 'out' else False
@@ -121,6 +177,7 @@ def add(request, reg_number):
 
         # Get all submissions for the current month for this user
         monthly_submissions = Laundry.objects.filter(
+            user=request.user,
             reg_number=reg_number,
             date__gte=first_day_of_month  # Submissions from the first day of the month
         )
